+++ src/main/java/com/nhnacademy/order_payments/dto/order/PackagingDto.java
@@ -1,12 +1,23 @@
+/*
+ * +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+ * + Copyright 2025. NHN Academy Corp. All rights reserved.
+ * + * While every precaution has been taken in the preparation of this resource,  assumes no
+ * + responsibility for errors or omissions, or for damages resulting from the use of the information
+ * + contained herein
+ * + No part of this resource may be reproduced, stored in a retrieval system, or transmitted, in any
+ * + form or by any means, electronic, mechanical, photocopying, recording, or otherwise, without the
+ * + prior written permission.
+ * +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+ */
+
 package com.nhnacademy.order_payments.dto.order;
 
 import com.nhnacademy.order_payments.entity.Packaging;
-import jakarta.persistence.Column;
 
 public record PackagingDto(
-        long id,
+        Long id,
         String name,
-        int price
+        Long price
 ) {
     public PackagingDto(Packaging packaging) {
         this(
